+++ app/src/main/java/com/fieldbook/tracker/traits/PercentTraitLayout.java
@@ -1,5 +1,8 @@
 package com.fieldbook.tracker.traits;
 
+import static com.fieldbook.tracker.traits.NumericTraitLayout.isOver;
+import static com.fieldbook.tracker.traits.NumericTraitLayout.isUnder;
+
 import android.app.Activity;
 import android.content.Context;
 import android.graphics.Color;
@@ -8,10 +11,15 @@ import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+
 import com.fieldbook.tracker.R;
 import com.fieldbook.tracker.activities.CollectActivity;
 import com.fieldbook.tracker.database.models.ObservationModel;
 import com.fieldbook.tracker.objects.TraitObject;
+import com.fieldbook.tracker.traits.formats.Formats;
+import com.fieldbook.tracker.traits.formats.Scannable;
+import com.fieldbook.tracker.traits.formats.TraitFormat;
 
 public class PercentTraitLayout extends BaseTraitLayout {
     private SeekBar seekBar;
@@ -60,7 +68,7 @@ public class PercentTraitLayout extends BaseTraitLayout {
 
         seekListener = new SeekBar.OnSeekBarChangeListener() {
 
-            public void onProgressChanged(SeekBar sb, int progress, boolean arg2) {
+            public void onProgressChanged(SeekBar sb, int progress, boolean fromUser) {
                 int minimum = 0;
                 if (getCurrentTrait() != null) {
                     try {
@@ -72,7 +80,13 @@ public class PercentTraitLayout extends BaseTraitLayout {
                 if (sb.getProgress() < minimum)
                     sb.setProgress(minimum);
 
-                setCurrentValueText(sb.getProgress(), Color.parseColor(getDisplayColor()));
+                getCollectInputView().setText(sb.getProgress() + "%");
+
+                // check if the change was from user interaction.
+                // useful when navigating across repeated values
+                if (fromUser) {
+                    setCurrentValueAsEdited();
+                }
             }
 
             public void onStartTrackingTouch(SeekBar sb) {
@@ -131,7 +145,7 @@ public class PercentTraitLayout extends BaseTraitLayout {
                 getCollectInputView().setText("NA");
                 getSeekBar().setProgress(0);
             } else if (!model.getValue().isEmpty()) {
-                getSeekBar().setProgress(Integer.parseInt(model.getValue()));
+                setSeekBarProgress(model.getValue());
             } else {
                 super.refreshLayout(onNew);
             }
@@ -144,18 +158,16 @@ public class PercentTraitLayout extends BaseTraitLayout {
 
         if (value != null && !value.equals("NA") && !value.isEmpty()) {
 
-            seekBar.setMax(Integer.parseInt(getCurrentTrait().getMaximum()));
-
-            int textColor = value.equals(getDefaultValue()) ? Color.BLACK : Color.parseColor(getDisplayColor());
-            setCurrentValueText(value, textColor);
+            // Default to max 100 if maximum is not set
+            String maxString = getCurrentTrait().getMaximum();
+            setSeekBarMax(maxString);
 
             seekBar.setOnSeekBarChangeListener(null);
-            seekBar.setProgress(Integer.parseInt(value));
+            setSeekBarProgress(value);
             seekBar.setOnSeekBarChangeListener(seekListener);
 
         } else if (value != null && value.equals("NA")) {
             getCollectInputView().setText("NA");
-            getCollectInputView().setTextColor(Color.parseColor(getDisplayColor()));
             seekBar.setProgress(0);
         }
     }
@@ -170,7 +182,8 @@ public class PercentTraitLayout extends BaseTraitLayout {
     @Override
     public void afterLoadDefault(CollectActivity act) {
         super.afterLoadDefault(act);
-        updateLoadBarValue(getDefaultValue());
+        getCollectInputView().setText(getDefaultValue() + "%");
+        updateLoadBar();
     }
 
     @Override
@@ -184,19 +197,25 @@ public class PercentTraitLayout extends BaseTraitLayout {
         }
     }
 
+    @NonNull
+    @Override
+    public Boolean validate(String data) {
+        try {
+            TraitObject trait = getCurrentTrait();
+            return !(isUnder(trait, data) || isOver(trait, data));
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
     private void updateLoadBar() {
         String max = getCurrentTrait().getMaximum();
         if (!max.isEmpty()) {
             seekBar.setMax(Integer.parseInt(max));
         }
         seekBar.setOnSeekBarChangeListener(null);
-        seekBar.setProgress(Integer.parseInt(getDefaultValue()));
+        setSeekBarProgress(getDefaultValue());
         seekBar.setOnSeekBarChangeListener(seekListener);
-    }
-
-    private void updateLoadBarValue(String value) {
-        setCurrentValueText(value, Color.BLACK);
-        updateLoadBar();
     }
 
     private String getDefaultValue() {
@@ -208,33 +227,52 @@ public class PercentTraitLayout extends BaseTraitLayout {
         return defaultValue;
     }
 
-    private void setCurrentValueText(int value, int color) {
-        setCurrentValueText(String.valueOf(value), color);
-    }
-
-    private void setCurrentValueText(String value, int color) {
-        getCollectInputView().setTextColor(color);
-        if (value.isEmpty())
-            getCollectInputView().setText(value);
-        else
-            getCollectInputView().setText(value + "%");
-    }
-
     @Override
     public void deleteTraitListener() {
-        removeTrait(getCurrentTrait().getName());
+        removeTrait(getCurrentTrait());
         super.deleteTraitListener();
         ObservationModel model = getCurrentObservation();
         seekBar.setOnSeekBarChangeListener(null);
         if (model != null) {
-            setCurrentValueText(model.getValue(), Color.BLACK);
-            seekBar.setProgress(Integer.parseInt(model.getValue()));
+            getCollectInputView().setText(model.getValue() + "%");
+            if (model.getValue().equals("NA")) {
+                seekBar.setProgress(0);
+            } else {
+                setSeekBarProgress(model.getValue());
+            }
         } else {
             String defaultValue = getDefaultValue();
-            setCurrentValueText(defaultValue, Color.BLACK);
-            seekBar.setProgress(Integer.parseInt(defaultValue));
+            getCollectInputView().setText("");
+            setSeekBarProgress(defaultValue);
         }
 
         seekBar.setOnSeekBarChangeListener(seekListener);
+    }
+
+    private void setSeekBarProgress(String value) {
+
+        ((CollectActivity) controller.getContext()).runOnUiThread(() -> {
+            try {
+                seekBar.setProgress(Integer.parseInt(value));
+                return;
+            } catch (NumberFormatException e) {
+                e.printStackTrace();
+            }
+            seekBar.setProgress(0);
+        });
+    }
+
+    private void setSeekBarMax(String max) {
+        ((CollectActivity) controller.getContext()).runOnUiThread(() -> {
+            try {
+                if (max != null && !max.isEmpty()) {
+                    seekBar.setMax(Integer.parseInt(max));
+                }
+                return;
+            } catch (NumberFormatException e) {
+                e.printStackTrace();
+            }
+            seekBar.setMax(100);
+        });
     }
 }